docs(atomic-react): add examples for folding component
https://coveord.atlassian.net/browse/KIT-1622
