+++ packages/samples/atomic-react/src/HeaderLink.tsx
@@ -1,0 +1,23 @@
+import React, {FunctionComponent} from 'react';
+
+type Props = {
+  page: string;
+  currentPage: string;
+  setPage(page: string): void;
+};
+export const HeaderLink: FunctionComponent<Props> = ({
+  page,
+  currentPage,
+  setPage,
+}: Props) => {
+  if (currentPage === page) return null;
+  return (
+    <li
+      onClick={() => {
+        setPage(page);
+      }}
+    >
+      {page} example
+    </li>
+  );
+};
