+++ packages/samples/atomic-react/src/style.css
@@ -1,0 +1,22 @@
+header {
+  padding: 10px 20px;
+}
+
+header ul li {
+  display: inline-block;
+  list-decoration: none;
+  cursor: pointer;
+  color: var(--atomic-primary);
+  margin-right: 24px;
+}
+
+header ul {
+  display: inline-block;
+  font-size: var(--atomic-text-sm);
+  font-family: var(--atomic-font-family);
+}
+
+.pageTitle {
+  font-weight: var(--atomic-font-bold);
+  font-family: var(--atomic-font-family);
+}
